Added one-to-many and many-to-many relationships

+++ Control_Shift/controllers/keyboard_controller.py
@@ -1,7 +1,9 @@
-from flask import Blueprint, jsonify, request, render_template, redirect, url_for
+from flask import Blueprint, jsonify, request, render_template, redirect, url_for, abort, current_app
 from main import db
 from models.keyboards import Keyboard
 from schemas.keyboard_schema import keyboard_schema, keyboards_schema
+from flask_login import login_required, current_user
+import boto3
 
 keyboards = Blueprint('keyboards', __name__)
 
@@ -26,8 +28,10 @@ def get_keyboards():
 
 # The POST route endpoint
 @keyboards.route("/keyboards/", methods=["POST"])
+@login_required
 def create_keeb():
     new_keeb = keyboard_schema.load(request.form)
+    new_keeb.creator = current_user
     db.session.add(new_keeb)
     db.session.commit()
     return redirect(url_for("keyboards.get_keyboards"))
@@ -36,17 +40,33 @@ def create_keeb():
 @keyboards.route("/keyboards/<int:id>/", methods=["GET"])
 def get_keyboard(id):
     keyboard = Keyboard.query.get_or_404(id)
+
+    s3_client=boto3.client("s3")
+    bucket_name=current_app.config["AWS_S3_BUCKET"]
+    image_url = s3_client.generate_presigned_url(
+        'get_object',
+        Params={
+            "Bucket": bucket_name,
+            "Key": keyboard.image_filename
+        },
+        ExpiresIn=100
+    )
+
     data = {
         "page_title": "Keyboard Detail",
-        "keyboard": keyboard_schema. dump(keyboard)
+        "keyboard": keyboard_schema.dump(keyboard),
+        "image": image_url
     }
     return render_template("keyboard_detail.html", page_data=data)
 
 
 @keyboards.route("/keyboards/<int:id>/", methods=["POST"])
+@login_required
 def update_keyboard(id):
     keyboard = Keyboard.query.filter_by(keyboard_id=id)
     
+    if current_user.id != keyboard.first().creator_id:
+        abort(403, "You do not have permission to alter this keyboard.")
 
     updated_fields = keyboard_schema.dump(request.form)
     if updated_fields:
@@ -59,10 +79,31 @@ def update_keyboard(id):
     }
     return render_template("keyboard_index.html", page_data=data)
 
+@keyboards.route("/keyboards/<int:id>/enrol/", methods=["POST"])
+@login_required
+def enrol_keyboard(id):
+    keyboard = Keyboard.query.get_or_404(id)
+    keyboard.owners.append(current_user)
+    db.session.commit()
+    return redirect(url_for('users.user_detail'))
 
-@keyboards.route("/keyboards/<int:id>/delete", methods=["POST"])
+@keyboards.route("/keyboards/<int:id>/drop/", methods=["POST"])
+@login_required
+def drop_keyboard(id):
+    keyboard = Keyboard.query.get_or_404(id)
+    keyboard.owners.remove(current_user)
+    db.session.commit()
+    return redirect(url_for("users.user_detail"))
+
+
+@keyboards.route("/keyboards/<int:id>/delete/", methods=["POST"])
+@login_required
 def delete_keyboard(id):
     keyboard = Keyboard.query.get_or_404(id)
+
+    if current_user.id != keyboard.creator_id:
+        abort(403, "You do not have persmission to delete this keyboard.")
+        
     db.session.delete(keyboard)
     db.session.commit()
     return redirect(url_for("keyboards.get_keyboards"))
